+++ nn_dataflow/core/loop_blocking_scheme.py
@@ -107,6 +107,7 @@ class LoopBlockingScheme(object):
 
         self.lcnt = util.prod(bl_tp)
 
+        # Need to define time for invalid scheme.
         self.time = float('inf')
 
         # Buffer data size for one unit.
@@ -174,6 +175,10 @@ class LoopBlockingScheme(object):
             self.valid = False
             return
 
+        # Array bus.
+        self.array_bus_width = resource.array_bus_width
+        # DRAM bandwidth.
+        self.dram_bandwidth = resource.dram_bandwidth
         # Parallel partitioning.
         self.num_nodes = resource.proc_region.dim.size()
         # Occupation.
@@ -184,6 +189,9 @@ class LoopBlockingScheme(object):
         self.finalized_stats = False
         self.ops = float('nan')
         self.time = float('nan')
+        self.proc_time = float('nan')
+        self.bus_time = float('nan')
+        self.dram_time = float('nan')
         self.access = [[float('nan')] * de.NUM for _ in range(me.NUM)]
         # Remote gbuf access.
         self.remote_gbuf_access = [0.] * de.NUM
@@ -275,6 +283,9 @@ class LoopBlockingScheme(object):
         return OrderedDict([('cost', self.get_cost(cost)),
                             ('ops', self.ops),
                             ('time', self.time),
+                            ('proc_time', self.proc_time),
+                            ('bus_time', self.bus_time),
+                            ('dram_time', self.dram_time),
                             ('access', self.access),
                             ('fetch', self.fetch),
                             ('remote_gbuf_access', self.remote_gbuf_access),
@@ -438,7 +449,7 @@ class LoopBlockingScheme(object):
 
         self.ops = self.nld.unit_ops * self.lcnt * self.num_nodes \
                 * self.part_occ
-        self.time = self.nld.unit_time * self.lcnt
+        self.proc_time = self.nld.unit_time * self.lcnt
 
         self.access[me.REGF] = [v * self.lcnt * t
                                 * self.num_nodes * self.part_occ for v, t
@@ -469,6 +480,16 @@ class LoopBlockingScheme(object):
         if not self.dst_is_dram:
             self.remote_gbuf_access[de.OFM] += self.access[me.DRAM][de.OFM]
             self.access[me.DRAM][de.OFM] = 0
+
+        # DRAM access time.
+        self.dram_time = sum(self.access[me.DRAM]) / self.dram_bandwidth
+
+        # Array multicast uses separate bus for each data category.
+        # Each data from GBUF takes one cycle to multicast to PEs.
+        self.bus_time = util.idivc(max(self.access[me.GBUF]) // self.num_nodes,
+                                   self.array_bus_width)
+
+        self.time = max(self.proc_time + self.bus_time, self.dram_time)
 
         self.finalized_stats = True