+++ nn_dataflow/core/pipeline_segment.py
@@ -429,12 +429,19 @@ class PipelineSegment(object):
                     dst_data_region = proc_region
 
                 # Make resource.
+                # Note that DRAM bandwidth is not split here. We optimistically
+                # assume each layer can use the full DRAM bandwidth at
+                # different time. We adjust this assumption when calculating
+                # the segment timing.
                 rtpl += (Resource(proc_region=proc_region,
                                   data_regions=(src_data_region,
                                                 dst_data_region),
                                   dim_array=self.resource.dim_array,
                                   size_gbuf=self.resource.size_gbuf,
-                                  size_regf=self.resource.size_regf),)
+                                  size_regf=self.resource.size_regf,
+                                  array_bus_width=self.resource.array_bus_width,
+                                  dram_bandwidth=self.resource.dram_bandwidth,
+                                 ),)
 
             assert len(rtpl) == len(ltpl)
             self.alloc += (rtpl,)