+++ nn_dataflow/tests/map_strategy_test/test_map_strategy_fixture.py
@@ -69,5 +69,6 @@ class TestMapStrategyFixture(unittest.TestCase):
         # Eyeriss, ISSCC'16, JSSC'17.
         self.resource['BASE'] = Resource(
             proc_region=proc_region, data_regions=data_regions,
-            dim_array=PhyDim2(12, 14), size_gbuf=108*1024, size_regf=520)
+            dim_array=PhyDim2(12, 14), size_gbuf=108*1024, size_regf=520,
+            array_bus_width=float('inf'), dram_bandwidth=float('inf'))