+++ nn_dataflow/tests/partition_test/test_partition_fixture.py
@@ -74,6 +74,11 @@ class TestPartitionFixture(unittest.TestCase):
                                        partition_batch=True,
                                        partition_ifmaps=False,
                                        **optdict)
+        self.options['DIMPART'] = Option(partition_hybrid=True,
+                                         partition_batch=True,
+                                         partition_ifmaps=True,
+                                         partition_dimensional=True,
+                                         **optdict)
 
     def _gen_partition(self, wlkey='BASE', dnkey='BASE', optkey='BASE',
                        guaranteed=False):