+++ nn_dataflow/tests/unit_test/test_nn_dataflow_scheme.py
@@ -34,6 +34,7 @@ from nn_dataflow.core import SchedulingResult
 
 class TestNNDataflowScheme(unittest.TestCase):
     ''' Tests for NNDataflowScheme. '''
+    # pylint: disable=too-many-public-methods
 
     def setUp(self):
         self.network = Network('test_net')
@@ -52,7 +53,9 @@ class TestNNDataflowScheme(unittest.TestCase):
                        type=NodeRegion.DATA))
 
         self.c1res = SchedulingResult(
-            dict_loop=OrderedDict([('cost', 1.), ('time', 200), ('ops', 4.),
+            dict_loop=OrderedDict([('cost', 1.), ('time', 200.), ('ops', 4.),
+                                   ('proc_time', 200), ('bus_time', 0),
+                                   ('dram_time', 0),
                                    ('access', [[7, 8, 9]] * me.NUM),
                                    ('ti', [2, 2, 3]),
                                    ('to', [1, 2, 3]),
@@ -72,6 +75,8 @@ class TestNNDataflowScheme(unittest.TestCase):
 
         self.p1res = SchedulingResult(
             dict_loop=OrderedDict([('cost', 0.1), ('time', 5), ('ops', 0.1),
+                                   ('proc_time', 5), ('bus_time', 0),
+                                   ('dram_time', 0),
                                    ('access', [[.7, .8, .9]] * me.NUM),
                                    ('ti', [2, 2, 3]),
                                    ('to', [1, 2, 3]),
@@ -227,7 +232,6 @@ class TestNNDataflowScheme(unittest.TestCase):
             self.assertAlmostEqual(a, (7 + 8 + 9) + (.7 + .8 + .9) * 2)
         self.assertAlmostEqual(self.dtfl.total_noc_hops,
                                (4 + 5 + 6) + (.4 + .5 + .6) * 2)
-        self.assertAlmostEqual(self.dtfl.total_node_time, 200 * 4 + 5 * 2 * 2)
 
     def test_time_full_net_single_seg(self):
         ''' time() when full network fits in a single segment. '''
@@ -245,6 +249,13 @@ class TestNNDataflowScheme(unittest.TestCase):
         dtfl['p1'] = self.p1res
         dtfl['p2'] = self.p2res._replace(sched_seq=(1, 0, 0))
         self.assertListEqual(dtfl.segment_time_list(), [205, 5])
+
+    def test_total_static_cost(self):
+        ''' Static cost. '''
+        self.assertAlmostEqual(self.dtfl.total_static_cost(1),
+                               200 * 4 + 5 * 2 * 2)
+        self.assertAlmostEqual(self.dtfl.total_static_cost(2),
+                               2 * self.dtfl.total_static_cost(1))
 
     def test_stats_active_node_pes(self):
         ''' Per-layer stats: active node PEs. '''